requirements.txt
certifi==2019.11.28
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.18.1
Pillow==7.0.0
scipy==1.4.1
six==1.14.0
Werkzeug==0.16.0
